+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - ✨(back) add celery for running background tasks
 - 🧱(helm) add celery worker and beat deployments
 - ✨(back) process project files asynchronously with indexing state
+- ✨(back) process conversation files asynchronously on a worker
 
 ## [0.0.19] - 2026-06-24